refactor multiplayer manager -> data sender
