+++ TurboMinerVirtualTable.Unity/Assets/Scripts/Networking/DataSender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ebaeb9e4550ef14388cc90fa52bf6e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
